remove confusing usage of slot list
while reading the code, i found the usage of the slot list rather
confusing. the intent was to optimize looking up of a free job slot
(index into the job array), instead of iterating over the job array
to find an unused entry (pid == -1). however this can be entirely
left away if we just fill the job array sequentially till it's full,
and later by waiting for/reaping a single process and re-use its
job slot immediately when we want to add a job.