add new -pipe command line option
sometimes the called program wants input on stdin, and while it is easily
possible to write a short shell script wrapper that takes an argument and
passes it to the stdin of that program, it's more convenient to pass input
directly. and in some cases it is critical that no cpu time gets lost by
launching the extra wrapper program.