add EOF marker functionality
since it's hard or impossible to close fifo pipes in shell scripts,
it's handy to have a way to signal jobflow when the end of input
is reached so it can terminate once everything is processed.
the shell script can now run jobflow with -eof=XXX to specify a
magic eof marker.