bump version to 1.2.4
this is mainly to bundle a release with latest libulz changes to optparser
which makes it possible to use e.g. --foo bar additional to --foo=bar.
this is mainly to make -j 16 work, since it took myself several times by
surprise that this didn't seem to work, so i guess the same issue would
strike other users too.