propagate error exit status of called process
previously jobflow ignored when a process returned error exit
status (plus it failed to recognize abnormal process termination
of a child, like SIGSEGV and went into an endless loop in such a
case).

the new behaviour is that when a child process fails, execution
stops and jobflow returns error exit status (although not the same,
as different child processes may return different error status).
the returned error status is always 1.